requirements.txt
streamlit>=1.37
pandas>=2.1
numpy>=1.26
openpyxl>=3.1
xlsxwriter>=3.2
xlrd>=2.0.1
pyxlsb>=1.0.10
odfpy>=1.4.1
beautifulsoup4>=4.12
lxml>=5.2
python-dateutil>=2.9
holidays>=0.60